Final Version
In this project I have customized the open source MBlock software to have some more blocks and features, acutllay I have removed some bugs, which where related to its wifi and bluetooth connection to have a more reliable connection. also in this version I added functions to control the threads which are running on system and under the program itself.